Arrange file in the folder for smooth execution
